Add files via upload
Transfer Token Smart Contract

This Solidity-based smart contract enables the owner to mint new tokens and transfer tokens between Ethereum addresses. It also allows the owner to retrieve the total supply of the tokens.

The key features of this smart contract include:

1. Token Minting: The contract owner can mint new tokens as per their requirement.

2. Token Transfer: The contract allows token holders to transfer their tokens to other Ethereum addresses.

3. Total Supply: The total number of tokens in circulation can be queried using the contract's totalSupply() function.

4. Owner Control: The contract owner maintains full control over the token minting process and can determine the total token supply.

This smart contract was developed and tested in the Remix IDE environment, a popular web-based Ethereum IDE. It provides a secure and customizable token management solution for decentralized applications built on the Ethereum blockchain.